Moved middlewares to their namespace
